argus_service/requirements.txt
flask==3.1.3
dnspython==2.8.0
